requirements.txt
alabaster==0.7.12
arrow==0.14.7
asn1crypto==0.24.0
Babel==2.9.1
caldav==0.6.2
certifi==2018.8.24
chardet==3.0.4
clang==6.0.0.2
cryptography==2.6.1
cycler==0.10.0
dateparser==1.0.0
Django==2.2.5
django-adminlte2==0.4.1
django-background-tasks==1.2.5
django-colorfield==0.4.5
django-compat==1.0.15
django-haystack==3.0
djangorestframework==3.10.3
docutils==0.17.1
entrypoints==0.3
ExifRead==2.1.2
fitparse==1.1.0
fpdf==1.7.2
future==0.18.2
geographiclib==1.50
geopy==1.20.0
gpsphoto==2.2.3
ics==0.7
idna==2.6
imagesize==1.2.0
importlib-metadata==3.4.0
Jinja2==3.0.1
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.2.0
lxml==4.4.2
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.2.2
mysqlclient==1.4.4
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.1.26
packaging==21.0
photocollage==1.4.4
piexif==1.1.3
Pillow==6.2.1
pybind11==2.4.3
pycairo==1.20.0
pyCardDAV==0.7.0
pycrypto==2.6.1
Pygments==2.10.0
PyGObject==3.30.4
PyMySQL==0.9.3
pyocclient==0.4
pyparsing==2.4.7
python-apt==1.8.4
python-dateutil==2.7.3
python-twitter==3.5
pytz==2019.2
pyxdg==0.25
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
scipy==1.4.1
SecretStorage==2.3.1
six==1.12.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.3.0
ssh-import-id==5.7
staticmap==0.5.5
TatSu==4.4.0
typing-extensions==3.7.4.3
tzlocal==2.0.0
urllib3==1.24.1
urwid==2.1.0
vcard==0.13.0
vobject==0.9.6.1
Whoosh==2.7.4
wordcloud==1.7.0
xmltodict==0.12.0
zipp==3.4.0
